+++ src/components/Header.jsx
@@ -1,0 +1,7 @@
+const Header = (props) => {
+    return (
+        <h1>{props.course}</h1>
+    )
+}
+
+export default Header
