+++ src/components/Part.jsx
@@ -1,0 +1,9 @@
+const Part = ({part}) => {
+    return (
+        <div>
+            <p> {part.name} {part.exercises}</p>
+        </div>
+    )
+}
+
+export default Part
